Global ref module connection
